Refactor matchers and remove redundant backticks in InputAudio.
Moved `OpenAiMatchers` to `completions` package and renamed it for better organization. Updated references accordingly. Replaced unnecessary backticks in `InputAudio` to improve code readability and maintain consistency.
