+++ ai-mocks-openai/src/commonMain/kotlin/me/kpavlov/aimocks/openai/completions/OpenaiCompletionsMatchers.kt
@@ -1,14 +1,15 @@
-package me.kpavlov.aimocks.openai
+package me.kpavlov.aimocks.openai.completions
 
 import io.kotest.matchers.Matcher
 import io.kotest.matchers.MatcherResult
+import me.kpavlov.aimocks.openai.ChatCompletionRequest
 import me.kpavlov.aimocks.openai.model.ChatCompletionRole
 
-internal object OpenAiMatchers {
+internal object OpenaiCompletionsMatchers {
     fun systemMessageContains(string: String): Matcher<ChatCompletionRequest?> =
         object : Matcher<ChatCompletionRequest?> {
             override fun test(value: ChatCompletionRequest?): MatcherResult =
-                MatcherResult(
+                MatcherResult.Companion(
                     value != null &&
                         value.messages
                             .find {
@@ -26,7 +27,7 @@ internal object OpenAiMatchers {
     fun userMessageContains(string: String): Matcher<ChatCompletionRequest?> =
         object : Matcher<ChatCompletionRequest?> {
             override fun test(value: ChatCompletionRequest?): MatcherResult =
-                MatcherResult(
+                MatcherResult.Companion(
                     value != null &&
                         value.messages
                             .find { it.role == ChatCompletionRole.USER }
